JDK-8261447: MethodInvocationCounters frequently run into overflow

+++ src/hotspot/share/runtime/sharedRuntime.hpp
@@ -565,16 +565,16 @@ class SharedRuntime: AllStatic {
 
   // Statistics code
   // stats for "normal" compiled calls (non-interface)
-  static int     _nof_normal_calls;              // total # of calls
-  static int     _nof_optimized_calls;           // total # of statically-bound calls
-  static int     _nof_inlined_calls;             // total # of inlined normal calls
-  static int     _nof_static_calls;              // total # of calls to static methods or super methods (invokespecial)
-  static int     _nof_inlined_static_calls;      // total # of inlined static calls
+  static int64_t _nof_normal_calls;               // total # of calls
+  static int64_t _nof_optimized_calls;            // total # of statically-bound calls
+  static int64_t _nof_inlined_calls;              // total # of inlined normal calls
+  static int64_t _nof_static_calls;               // total # of calls to static methods or super methods (invokespecial)
+  static int64_t _nof_inlined_static_calls;       // total # of inlined static calls
   // stats for compiled interface calls
-  static int     _nof_interface_calls;           // total # of compiled calls
-  static int     _nof_optimized_interface_calls; // total # of statically-bound interface calls
-  static int     _nof_inlined_interface_calls;   // total # of inlined interface calls
-  static int     _nof_megamorphic_interface_calls;// total # of megamorphic interface calls
+  static int64_t _nof_interface_calls;            // total # of compiled calls
+  static int64_t _nof_optimized_interface_calls;  // total # of statically-bound interface calls
+  static int64_t _nof_inlined_interface_calls;    // total # of inlined interface calls
+  static int64_t _nof_megamorphic_interface_calls;// total # of megamorphic interface calls
 
  public: // for compiler
   static address nof_normal_calls_addr()                { return (address)&_nof_normal_calls; }
@@ -586,7 +586,7 @@ class SharedRuntime: AllStatic {
   static address nof_optimized_interface_calls_addr()   { return (address)&_nof_optimized_interface_calls; }
   static address nof_inlined_interface_calls_addr()     { return (address)&_nof_inlined_interface_calls; }
   static address nof_megamorphic_interface_calls_addr() { return (address)&_nof_megamorphic_interface_calls; }
-  static void print_call_statistics(int comp_total);
+  static void print_call_statistics(uint64_t comp_total);
   static void print_statistics();
   static void print_ic_miss_histogram();
